+++ packages/data-model/src/converter/index.ts
@@ -1,2 +1,3 @@
 export * from './AcDbBatchProcessing'
 export * from './AcDbDxfConverter'
+export * from './worker'