+++ packages/libredwg-converter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mlightcad/libredwg-converter
 
+## 3.0.5
+
+### Patch Changes
+
+- feat: enable web worker for libredwg-converter
+- Updated dependencies
+  - @mlightcad/data-model@1.2.5
+
 ## 3.0.4
 
 ### Patch Changes